+++ StardewArchipelago/Locations/CodeInjections/Initializers/ModCodeInjectionInitializer.cs
@@ -18,14 +18,6 @@ namespace StardewArchipelago.Locations.CodeInjections.Initializers
         {
             _archipelago = archipelago;
             InitializeModdedContent(monitor, modHelper, archipelago, locationChecker, shopReplacer, shopStockGenerator);
-            // Fix to remove dupes in Railroad Boulder
-            var railroadBoulderOrder = SpecialOrder.GetSpecialOrder("Clint2", null);
-            var railroadDupeCount = Game1.player.team.specialOrders.Count(x => x.questKey.Value.Equals("Clint2Again"));
-            while (railroadDupeCount > 1)
-            {
-                Game1.player.team.specialOrders.Remove(railroadBoulderOrder);
-                railroadDupeCount -= 1;
-            }
         }
 
         private static void InitializeModdedContent(IMonitor monitor, IModHelper modHelper, ArchipelagoClient archipelago, LocationChecker locationChecker, ShopReplacer shopReplacer, ShopStockGenerator shopStockGenerator)
